Transformation: Remove obsolete RemoveCallTrafo and rename to drhook
