+++ transformations/tests/test_drhook.py
@@ -9,12 +9,12 @@ import shutil
 import pytest
 
 from loki import (
-    Scheduler, SFilter, ProcedureItem, SchedulerConfig, FindNodes, CallStatement,
-    gettempdir, OMNI, Import, Sourcefile
+    Scheduler, SFilter, ProcedureItem, SchedulerConfig, FindNodes,
+    CallStatement, gettempdir, OMNI, Import
 )
 
 from conftest import available_frontends
-from transformations import DrHookTransformation, RemoveCallsTransformation
+from transformations import DrHookTransformation
 
 
 @pytest.fixture(scope='module', name='config')
@@ -197,45 +197,3 @@ def test_dr_hook_transformation_remove(frontend, config, source):
             assert not drhook_calls
             assert not drhook_imports
             assert 'zhook_handle' not in item.ir.variables
-
-
-@pytest.mark.parametrize('include_intrinsics', (True, False))
-@pytest.mark.parametrize('kernel_only', (True, False))
-@pytest.mark.parametrize('frontend', available_frontends(
-    xfail=[(OMNI, 'Incomplete source tree impossible with OMNI')]
-))
-def test_utility_routine_removal(frontend, config, source, include_intrinsics, kernel_only):
-    """
-    Test removal of utility calls and intrinsics with custom patterns.
-    """
-    scheduler_config = SchedulerConfig.from_dict(config)
-    scheduler = Scheduler(paths=source, config=scheduler_config, frontend=frontend)
-    scheduler.process(
-        transformation=RemoveCallsTransformation(
-            routines=['ABOR1', 'WRITE(NULOUT', 'DR_HOOK'],
-            include_intrinsics=include_intrinsics, kernel_only=kernel_only
-        )
-    )
-
-    routine = scheduler['rick_rolled#never_gonna_give'].ir
-    transformed = routine.to_fortran()
-    assert '[SUBROUTINE CALL]' not in transformed
-    assert '[INLINE CONDITIONAL]' not in transformed
-    assert ('dave' not in transformed) == include_intrinsics
-    assert ('[WRITE INTRINSIC]' not in transformed) == include_intrinsics
-    assert 'zhook_handle' not in routine.variables
-
-    for r in routine.members:
-        transformed = r.to_fortran()
-        assert '[SUBROUTINE CALL]' not in transformed
-        assert '[INLINE CONDITIONAL]' not in transformed
-        assert ('dave' not in transformed) == include_intrinsics
-        assert 'zhook_handle' not in routine.variables
-
-    routine = Sourcefile.from_file(source/'never_gonna_give.F90', frontend=frontend)['i_hope_you_havent_let_me_down']
-    assert 'zhook_handle' in routine.variables
-    assert len([call for call in FindNodes(CallStatement).visit(routine.body) if call.name == 'dr_hook']) == 2
-
-    driver = scheduler['#rick_astley'].ir
-    drhook_calls = [call for call in FindNodes(CallStatement).visit(driver.body) if call.name == 'dr_hook']
-    assert len(drhook_calls) == (2 if kernel_only else 0)
