+++ transformations/transformations/__init__.py
@@ -16,7 +16,7 @@ from transformations.single_column_claw import * # noqa
 from transformations.single_column_coalesced import * # noqa
 from transformations.single_column_coalesced_vector import * # noqa
 from transformations.single_column_hoist import * # noqa
-from transformations.utility_routines import * # noqa
+from transformations.drhook import * # noqa
 from transformations.scc_cuf import * # noqa
 from transformations.pool_allocator import * # noqa
